--- tests/unit/services/mobile-menu-test.js
@@ -1,12 +1,0 @@
-import { moduleFor, test } from 'ember-qunit';
-
-moduleFor('service:mobile-menu', 'Unit | Service | mobile menu', {
-  // Specify the other units that are required for this test.
-  // needs: ['service:foo']
-});
-
-// Replace this with your real tests.
-test('it exists', function(assert) {
-  let service = this.subject();
-  assert.ok(service);
-});
